+++ quint/src/builtin.ts
@@ -30,7 +30,8 @@ export function builtinDocs(gen: IdGenerator): ParseResult<Map<string, Documenta
   // Read file and remove windows line endings (\r) using `lf`
   const sourceCode = lf(readFileSync(path, 'utf8'))
 
-  return parsePhase1fromText(gen, sourceCode, path).map(phase1Data => produceDocs(phase1Data.modules[0]))
+  const { modules } = parsePhase1fromText(gen, sourceCode, path)
+  return produceDocs(modules[0])
 }
 
 // TODO: Move builtinSignatures() to this file and read it from builtin.qnt, see https://github.com/informalsystems/quint/issues/452