+++ quint/test/flattening/flattener.test.ts
@@ -21,11 +21,10 @@ describe('flattenModule', () => {
       '\n'
     )} } module C { ${thirdModuleDecls.join('\n')} }`
 
-    const parseResult = parse(idGenerator, 'fake_location', fake_path, quintModules)
-    if (parseResult.isLeft()) {
-      assert.fail('Failed to parse mocked up module')
-    }
-    const { modules, table } = parseResult.unwrap()
+    const { modules, table, errors } = parse(idGenerator, 'fake_location', fake_path, quintModules)
+
+    assert.isEmpty(errors, 'Failed to parse mocked up module')
+
     const modulesByName: Map<string, QuintModule> = new Map(modules.map(m => [m.name, m]))
 
     modules.forEach(m => {