+++ quint/test/ir/idRefresher.test.ts
@@ -8,7 +8,6 @@ import { analyzeModules } from '../../src/quintAnalyzer'
 import { collectIds } from '../util'
 import { QuintModule, isDef } from '../../src/ir/quintIr'
 import JSONbig from 'json-bigint'
-import { quintErrorToString } from '../../src'
 
 describe('generateFreshIds', () => {
   // Generate ids starting from 100, so that we can easily distinguish them from
@@ -29,15 +28,8 @@ describe('generateFreshIds', () => {
 
   const quintModules: string = `module A { ${defs.join('\n')} }`
 
-  const parseResult = parse(idGenerator, 'fake_location', fake_path, quintModules)
-  if (parseResult.isLeft()) {
-    it('should parse the mocked up module', () => {
-      assert.fail(
-        `Failed to parse mocked up module. Errors: ${parseResult.value.errors.map(quintErrorToString).join('\n')}`
-      )
-    })
-  }
-  const { modules, table, sourceMap } = parseResult.unwrap()
+  const { modules, table, sourceMap, errors } = parse(idGenerator, 'fake_location', fake_path, quintModules)
+  assert.isEmpty(errors, 'Failed to parse mocked up module')
   const [module] = modules
 
   const [analysisErrors, analysisOutput] = analyzeModules(table, modules)