+++ quint/test/util.ts
@@ -93,8 +93,7 @@ export function parseMockedModule(text: string): ParserPhase3 {
   const idGen = newIdGenerator()
   const fake_path: SourceLookupPath = { normalizedPath: 'fake_path', toSourceName: () => 'fake_path' }
   const parseResult = parse(idGen, 'fake_location', fake_path, text)
-  if (parseResult.isLeft()) {
-    assert.fail('Failed to parse mocked up module')
-  }
-  return parseResult.unwrap()
+
+  assert.isEmpty(parseResult.errors)
+  return parseResult
 }